AI Perception Hearing: Investigative Noise Response
Implemented behaviour (in BT) for 'investigative' behaviour.
After hearing a noise, the AI rotates to face the noise source.
In practice, this means that if an enemy attacks the player and another enemy is nearby but faced away from the player, they will heard the noise of the first enemy going to attack the player and will turn around, from which they might be able to see the player in their field of view.